housekeeping: Error Macros for parser
All the errors in `parser` are moved to macro invocation. This allows us
to change the Error types if required later on add logging etc. Also,
this avoids a lot of boilerplate code like `.to_string()` and `.clone()`
to be repeated.

TODO: tokenize errors are not converted to Macro invocation yet.