Semantically correct `expect_tokens` behavior
In the previous implementation of `expect_tokens` we were returning
`false` when something could potentially be an error condition. This
required caller to ensure, passed slice is at-least of as much length as
that of `checkers`. Semantically this didn't make sense. Fixing the
implementation allows one to call with `tokens[start..]` without doing
the tricks of calculating the size etc. Simplifies caller as well.