Fixes `parse_range_elements`
Parse Range elements was not properly parsing `Value` as specified in
the grammar and hence could potentially fail. Fixed that to support as
specified in the grammar.

Added the scaffolds for adding more test cases. Need to add a few more
test cases, but at-least all the right scaffolds are places.