Support for `FieldSpec`s
While handling Class Definition, we were dumping everything inside a
class definition into a string to start with. This is more structured
now with support for `TypeFieldSpec` and `FixedTypeValueFieldSpec. The
current plan is to only support these two `FieldSpec` types and
leave others for now. Note: We are not yet properly handling 'WITH
SYNTAX for Information Objects. It's a bit involved. We'll need to split
and splice a few tokens a bit.