Better support for Object Set and small fixes
An object set can be made up of individual objects, object references or
object set references. The parsing supports all three options. Right now
it is simply dumped into a 'Reference', but we should have a proper enum
for that.

Also fixed an issue with handling of `DEFAULT` value, where we were not
advancing the tokens consumed when `DEFAULT` was present.

We are able to parse now RANAP and NGAP specifications.