Fixed a couple of edge cases in `parse_type`
Found an error in `parse_reference_type` where if the input ends on a
token, you'd get an error while first trying to `parse` for the class
field (`UnexpectedEnd`), that would fail the parsing. Found while
parsing `S1AP`.

Also now accept `VisibleString` as a character string type.

While Matching a CLASS Reference, we need to account for presence of
numbers as well.

Now parsing of S1AP is complete too.