Added support for `ObjectAssignment`
Previously, we were calling an `ObjectAssignment` incorrectly as a
`ValueAssignment` (Since a 'CLASS' reference can always be a 'Type'
reference.). Now we are properly handling Object assignment.

Where the Object Assignment matches the `WITH SYNTAX` of the `CLASS` is
not verified here as the definition of the class may not be available at
this stage.