Skeleton for Resolver Code
A basic high level skeleton for Resolver Code. To resolve a
'definition' into something a compiler understands. We get a mutalbe
reference to the definition, resolve it and mark it as resolved. This
requires us to have the `Asn1Module` to which the definition belongs to
be mutable. We'd need to use Interior Mutability for that. Hence now we
are using `RefCell<Asn1Module>` inside the compiler's `modules`
`HashMap`.