Processing Parameterized and ObjectClass  definitions
To process parameterized and object class definitions inside the
resolver does not require much to be done. This is because, a
parameterized definition and/or object class will never be 'resolved'
per se, but would be used by other definitions to 'resolve' themselves.
What we really need is an access to the original definition when such a
resolution is required. Right now simply maintaining them in their
respective maps inside the compiler, with definition name as keys.