Topologically sorting 'definitions' in a module.
In order to resolve definitions within a module in proper order, we are
Topologically sorting the definitions, so that a dependency will be
resolved before it's dependents are resolved. Right now we are doing it
for type references and class references only. Eventually this would
need to be done for object and value references as well.

This works well for NGAP specs, for working well for RANAP and S1AP
specs, we'd need to process Parameterized References as well.