Handling of ParameterizedType in Reference
We have a very simple skeleton implementation for handling
ParameterizedType (on the RHS of an assignment). Now we are able to
'resolve' definitions for All the protocols. So we can call it a first
pass of 'resolution' complete. As of now this is quite basic, but enough
to build upon. As we really handle each of the types/sets/Values we'll
get closer to representation that can be used to generate the code.