housekeeping: Using a proper `Resolver` struct
In a number of `resolve_something` functions, we were passing reference
to multiple definitions (eg. `resolved_defs`, `parameterized_defs`,
`object_classes` etc.) For every internal function that implemented the
`resolve_something` function, this whole baggage had to be passed.
Wrapped this in a nice `Resolver` function and maintaining it inside the
main `Compiler` struct. This cleans up a lot of code.

A side effect of that is - we could easily get a `mutable` reference to
modules and hence we no longer need to wrap it in a `RefCell`, since our
mutable borrows are more fine grained now and not the entire `Compiler`
structure.