Better handling of Referenced Types and Values refering them.
Referenced Types and values corresponding to Referenced Types are
handled correctly now. For that an additional Variant is defined on
`Asn1ResolvedType` called `Reference` that references a
`Asn1ResolvedDefinition`. Also, the corresponding `Value` definition
should also have it's `typeref` as a `Reference` to be explicit. The
`Reference`d value in turn will eventually have one of the `Base`
values.