Better handling of Object and ObjectSet
In the Parser, we were simply dumping the ObjectSet and Object as
String without attaching any semantic meaning to them. Now an ObjectSet
is made up of ObjectSetElements each of which Can be a reference to an
Object or ObjectSet or an Object from ObjectDefinition.

These additional references will need to be considered during dependency
resolution of the definitions. Thus defined `dependent_references`
method now on an ObjectSet (Not Defined yet on Object - Because We are
not Treating ObjectDefinition which can have an ObjectReference. This
may need to be handled perhaps later.)

Similarly, we have to handle it in the Resolver to get proper Resolved
definitions.

The actual Object definition is still not yet verified against the
'CLASS' definition and parsed, but that's next after handling the `WITH
SYNTAX`.