Processing Objects and Object Sets before Resolving
Previously, we were processing Objects and ObjectSets during 'resolve'
phase. This made a certain things messy. We needed to wait to resolve
the Object and ObjectSet values before all other definitions are
'resolved'. This in turn is a problem because certain Type definitions
(those using Table Constraints say.) could be dependent on these Object
Set definitions.

The correct time to process Objects (ie. Make sure they are in
accordance with their 'Class' definition.) is after we collect the
modules (so we know that the CLASS definition referred by an Object is
resolved), but before actual 'resolution' takes place.

We'd need to handle 'dependent' Types from these definitions and also
ObjectSet dependencies in certain Type definitions (Parameterized
Types). We can get started with that.