`resolve_constructed_type` returns `Asn1ResolvedType`
During resolution of constructed types (sequence, choice), it's possible
that the choice gets resolved to multiple values each of which is a
constructed type (due to ClassField references resolving to different
values in Set). To handle this, it's better to return the actual
Asn1ResolvedType and let the individual function handle it internally.