Definitions Resolutions 1st Pass Complete.
Now we are able to resolve definitions including those that are defined
using Object Sets. This should be good enough as a first pass
implementaton for the definition Resolution, Enough to actually generate
structures for the types. There are a few parts that are not handled
completely yet, like optional 'Sequence' Components etc. But this is
good enough as a first pass now.