Using `i128` as base type as opposed to `i64`
For some constants in S1AP/NGAP signed 64 was not sufficient (while
unsigned 64 was), but to make sure we can support negative numbers as
well, we are using the `i128` as base type. This is truly an overkill
for now. But When we actually determine the bitwidth, the generated code
won't be that 'wide'.