Fixed `resolve_integer` signature to return `Asn1ResolvedInteger`
While 'resolving' an integer, we were passing a mutable reference to a
default `Asn1ResolvedInteger` and modifying it, that's not very obvious
for the reader of the program, instead a default is allocated now in the
callee and the right type is returned. This also makes signature
consistent with other functions in parser and resolver.