Handling of Size Constraint for OCTET STRING And CharacterString
For OCTET STRING and all CharacterString types, in their `Resolved`
versions, the Size Constraint is handled as well. For now this is not
going to be used for Structs generation, but instead will be used when
we implements Codec and Validating Constraints.

Defined an API on Asn1Constraint `is_size_constraint`.