A couple of changes to Data for Consistent Code Generation
In the `Resolver`, `resolved_defs` were maintained in a HashMap, which
made the look up fast during the resolution, but while generating the
code, the keys in the `HashMap` would get ordered in different order
every time. This resulted in different code generation every time. This
is a bit annoying, especially when you've to search for the definitions.
Now instead the data is kept in a `BTreeMap`, this increases a lookup
cost a bit during resolution, but generated code is consistent (and will
match sha256 checksum from previous run.)

Previously we were using a HashMap in `Asn1ResolvedEnumerated`, while
generating the code, this would generate code in a Random Order (which
changes every time the code is generated, thanks to Rust's `HashMap`
randomizing the entries to avoid Hash Attacks. Converted it to a Vec of
tuples. As it is we were not using the Lookup, so perhaps this is a
correct representation.