Rust Structs and Enums generation first pass
For all the specs we are generating the Rust structures and Enums. This
completes the first pass of code generation. There are a few things that
need to be done properly - eg. We need to maintain the sequence of `IEs`
as per the order in which Objects are defined in specifications. But we
can start looking at generating the codecs. Focus will be to generate
`aper` codec for the structs/enums. First just the decoders.