Added a `decoder_ty` member to help in decoding
While decoding an Open Type, we need to 'know' which variant of the Enum
is to be decoded. To do this we need to be able to decode the type that
is to be used for decoding the rest of the Open Type. Added this in the
`ResolvedSetType` which may be used to decode the actual type.