Adding attributes for the `CHOICE` type.
Overall strategy for the decoder is as follows - Looking at the
attributes on the `struct` (every ASN type is a struct, including
INTEGER, ENUMERATED etc.) we decide what needs to be decoded, so for the
choice types, we will decode choice type (Which is basically decoding
the integer,) and then depending upon the value of the decoded choice
type, the variant will be decoded. We'll need to decode the next struct.

Also fixed a BUG in the `parser` that was incorrectly adding `additions`
in CHOICE to `root_components` when the `ChoiceAdditionGroup` was
without a `version`. We are now able to generate the attributes for the
CHOICE type.