Reorganization of `decode` function and other fixes
In the `AperCodec`, the `decode` function was a bound method. This
required the implementing type to have a `default` implementation. But
that's not an easy thing to do for `enum` types. Hence now we have an
associated type and then return the associated type from the decode
(which itself can be `Self`).

All the `CharacterString` types were combined into a single
CharacterString type, Now added the actual ASN.1 Type name. This 'may'
be used by the codecs.