Fixed an issue with Code Generation in Object Sets
In an Object Set, it is possbile that two different IDs have same Type.
This caused an issue where one field/variant was dropped since our Map
would then overwrite the previous entry. This was fixed by making sure
that the key to the map is (Key, TypeId) so that it is not duplicated.