Fixed Unused fields in various internal structs
Since we do not fully support all the features yet, certain fields of
the structs are reported as unused as warning by the clippy lint
`dead_code`. Marking all such fields as ignorable by renaming them. eg.
A field `id` that is unused is renamed to `_id`. When the proper support
will be added this will be renamed to `id` again (say).