Fixes an issue with Derive Macro for OPEN Types
There was an issue with the Derive Macro of OPEN type that would cause
incorrect encoding. This was related to the handling of the length of
encoding an open type (which was missed in the original implementation).
With this fix, now we are able to perform a roundtrip
decode/encode/decode of a string and verify that the strings and types
are identical.

This should fix #7.