Fixes #19 : handling 2's complement encode/decode
In the original decode for handling `uncosntrained_whole_number` we were
simply treating the encoded integer as `i128`. This would work on
positive integers, but would fail for negative integers.

Also, fixed an issue with `encode_unconstrained_whole_number` where we
were always using 16 bytes for negative integer, which was inconsistent
with the specification.

Also Added test cases for integers that we may have to decode.