Changes to `Asn1Compiler` API and initial `build.rs` support.
Changed the `Asn1Compiler` API to have a single `compile_files` API that
compiles the passed ASN.1 files and generates the code. This way User
does not have to know internal details.

Also started with `build.rs` support with an example `build.rs` file
inside the `examples` directory. This will be used to generate the
bindings during `cargo test`, so that we now remove a lot of 'manual'
compile copy/paste of the generated modules.