Infrastructure for writing tests for compiler
Initially there was only one API `compile_files` in the `Asn1Compiler`.
This is generally good for using the compiler with ASN.1 files, however
it's perhaps desirable to have the ability to simply `compile` a module
string. Added an infrastrucutre to do that and a test case for failing
compilation (#49).

Removed stray prints from the previous fix of constraints.