UPER Codec Support: Initial commit
The top level `aper` module inside `codecs` is renamed to `per` module
and the `aper` and `uper` modules will be submodules. Correspondingly
renamed the structure `AperCodecData` as `PerCodecData`.

`PerCodecData` and `Error` are exported from the `aper` module as
`AperCodecData` and `AperCodecError` respectively to keep the external
APIs still the same.

Renamed the `AperCodec` trait methods as `aper_encode` and `aper_decode`
respectively as each codec would have it's own `encode`/`decode`
functions, it's a good idea to have it properly named.