Visibility support for consts and new crate version
For the constants defined from the specs, the visibility support was
missing, which made all the constants private and hence could not be
used. We support visibility for generated structure. Extended the same
support for the constants.

Making a new crate version as this change can be used from the crate
version with this support.