asn-compiler (#88): Constraints and dependent refs
Added support for missing `WITH COMPONENTS` constraint. Currently, only
the constraint is parsed, the code generation does not deal with this
constraint. This will have to be handled at the application level.

For `SEQUENCE OF` types, the dependent references on the `SIZE`
constraints were not considered, this caused resolution error. Fixed
that.

Added a test case for parsing `WITH COMPONENTS` and changed the
`compile_string` API to take a param `parse_only` to test for correct
parsing. This API currently is used only by tests.