Correctly handling sequence additions
Sequences Additions (those after `...` Extension Marker) were treated as
components from the root components. This resulted in structures with
components being treated as root components and caused errors in codecs
(due to incorrect number of components determined).

Fixed in the parser and resolver. Also defined a new `enum` type to deal
with the fact that the sequence additions can be addition groups or
components (stand alone).